+++ DEnc/Commands/FFmpegCommandBuilder.cs
@@ -1,4 +1,5 @@
-﻿using DEnc.Models;
+﻿using DEnc.Commands.Interfaces;
+using DEnc.Models;
 using DEnc.Models.Interfaces;
 using DEnc.Serialization;
 using System;
@@ -76,7 +77,7 @@ namespace DEnc.Commands
                     bool copyThisStream = enableStreamCopying && quality.Bitrate == 0;
                     string path = Path.Combine(outputDirectory, $"{outputBaseFilename}_{(quality.Bitrate == 0 ? "original" : quality.Bitrate.ToString())}.mp4");
 
-                    IVideoCommandBuilder videoBuilder = VideoCommandBuilder.Initilize(video.index, quality.Bitrate, path, AdditionalVideoFlags);
+                    IFFmpegVideoCommandBuilder videoBuilder = FFmpegVideoCommandBuilder.Initilize(video.index, quality.Bitrate, path, AdditionalVideoFlags);
 
                     if (!copyThisStream)
                     {
@@ -110,7 +111,7 @@ namespace DEnc.Commands
 
             foreach (MediaStream audioStream in streams)
             {
-                IAudioCommandBuilder builder = AudioCommandBuilder.Initilize(audioStream, outputDirectory, outputBaseFilename, AdditionalAudioFlags);
+                IFFmpegAudioCommandBuilder builder = FFmpegAudioCommandBuilder.Initilize(audioStream, outputDirectory, outputBaseFilename, AdditionalAudioFlags);
                 StreamAudioFile streamFile = builder
                     .WithLanguage()
                     .WithTitle()
@@ -160,267 +161,5 @@ namespace DEnc.Commands
             }
             return this;
         }
-    }
-
-    internal class VideoCommandBuilder : IVideoCommandBuilder
-    {
-        int index;
-        int bitrate;
-        string path;
-        List<string> commands;
-
-        public static IVideoCommandBuilder Initilize(int index, int bitrate, string path, ICollection<string> additionalFlags)
-        {
-            IVideoCommandBuilder builder = new VideoCommandBuilder(index, bitrate, path, additionalFlags);
-            return builder;
-        }
-
-        private VideoCommandBuilder(int index, int bitrate, string path, ICollection<string> additionalFlags)
-        {
-            this.index = index;
-            this.bitrate = bitrate;
-            this.path = path;
-            commands = new List<string>();
-
-            commands.Add($"-map 0:{index}");
-            if(additionalFlags != null && additionalFlags.Any())
-            {
-                commands.AddRange(additionalFlags);
-            }            
-        }
-
-        public StreamVideoFile Build()
-        {
-            commands.Add($"\"{ path}\"");
-
-            return new StreamVideoFile
-            {
-                Type = StreamType.Video,
-                Index = index,
-                Bitrate = bitrate.ToString(),
-                Path = path,
-                Argument = string.Join(" ", commands)
-            };
-        }
-
-        public IVideoCommandBuilder WithSize(IQuality quality)
-        {
-            if (quality.Width == 0 || quality.Height == 0)
-            {
-                return this;
-            }
-            commands.Add($"-s {quality.Width}x{quality.Height}");
-            return this;
-        }
-
-        public IVideoCommandBuilder WithBitrate(int bitrate)
-        {
-            if (bitrate == 0)
-            {
-                return this;
-            }
-            commands.Add($"-b:v {bitrate}k");
-            return this;
-        }
-
-        public IVideoCommandBuilder WithBitrate(int bitrate, int defaultBitrate)
-        {
-            if (bitrate == 0)
-            {
-                commands.Add($"-b:v {defaultBitrate}k");
-                return this;
-            }
-            commands.Add($"-b:v {bitrate}k");
-            return this;
-        }
-
-        public IVideoCommandBuilder WithPreset(H264Preset preset)
-        {
-            TryAddSimpleCommand($"-preset {preset}", preset.ToString());
-            return this;
-        }
-
-        public IVideoCommandBuilder WithProfile(H264Profile profile)
-        {
-            TryAddSimpleCommand($"-profile:v {profile}", profile.ToString());
-            return this;
-        }
-
-        public IVideoCommandBuilder WithProfileLevel(string level)
-        {
-            TryAddSimpleCommand($"-level {level}", level);
-            return this;
-        }
-
-        public IVideoCommandBuilder WithPixelFormat(string format)
-        {
-            TryAddSimpleCommand($"-pix_fmt {format}", format);
-            return this;
-        }
-
-        public IVideoCommandBuilder WithFramerate(int framerate)
-        {
-            if(framerate == 0)
-            {
-                return this;
-            }
-            commands.Add($"-r {framerate}");
-            return this;
-        }
-
-        public IVideoCommandBuilder WithVideoCodec(string sourceCodec, int keyframeInterval, bool enableCopy)
-        {
-            string defaultCoding = $"-x264-params keyint={keyframeInterval}:scenecut=0";
-
-            //TODO: Remove inline ternary and squash the switch statement
-            switch (sourceCodec)
-            {
-                case "h264":
-                    commands.Add($"-vcodec {(enableCopy ? "copy" : "libx264")} {defaultCoding}");
-                    break;
-                default:
-                    commands.Add($"-vcodec libx264 {defaultCoding}");
-                    break;
-            }
-            return this;
-        }
-
-        /// <summary>
-        /// Adds a simple command if the <paramref name="value"/> is not empty
-        /// </summary>
-        /// <remarks>
-        /// The point of this method is to reduce duplicated string.IsNullOrEmpty checks for the same command types
-        /// </remarks>
-        /// <param name="command">The fully qualified command assuming value is not empty</param>
-        /// <param name="value">The value of the command, if this is empty the provided command is discarded</param>
-        private void TryAddSimpleCommand(string command, string value)
-        {
-            if (string.IsNullOrEmpty(value))
-            {
-                return;
-            }
-            commands.Add(command);
-        }
-    }
-
-    internal class AudioCommandBuilder : IAudioCommandBuilder
-    {
-        List<string> commands;
-
-        MediaStream audioStream;
-        string outputDirectory;
-        string outputBaseFilename;
-        bool codecSupported;
-
-        string path;
-        string language;
-        string title;
-
-        public static IAudioCommandBuilder Initilize(MediaStream audioStream, string outputDirectory, string outputBaseFilename, ICollection<string> additionalFlags)
-        {
-            IAudioCommandBuilder builder = new AudioCommandBuilder(audioStream, outputDirectory, outputBaseFilename, additionalFlags);
-            return builder;
-        }
-
-        private AudioCommandBuilder(MediaStream audioStream, string outputDirectory, string outputBaseFilename, ICollection<string> additionalFlags)
-        {
-            this.audioStream = audioStream;
-            this.outputDirectory = outputDirectory;
-            this.outputBaseFilename = outputBaseFilename;
-            commands = new List<string>();
-
-            codecSupported = Constants.SupportedCodecs.ContainsKey(audioStream.codec_name);
-
-            commands.Add($"-map 0:{audioStream.index}");
-            if (additionalFlags != null && additionalFlags.Any())
-            {
-                commands.AddRange(additionalFlags);
-            }
-        }
-
-        public StreamAudioFile Build()
-        {
-            path = Path.Combine(outputDirectory, $"{outputBaseFilename}_audio_{language}_{audioStream.index}.mp4");
-            commands.Add($"\"{path}\"");
-
-            return new StreamAudioFile
-            {
-                Type = StreamType.Audio,
-                Index = audioStream.index,
-                Name = $"{language} {title}",
-                Path = path,
-                Argument = string.Join(" ", commands)
-            };
-        }
-
-        public IAudioCommandBuilder WithLanguage()
-        {
-            string language = audioStream.tag
-                .Where(x => x.key == "language")
-                .Select(x => x.value)
-                .FirstOrDefault();
-
-            if(language is null)
-            {
-                language = audioStream.disposition.@default > 0 ? "default" : "und";
-            }
-            this.language = language;
-            return this;
-        }
-
-        public IAudioCommandBuilder WithTitle()
-        {
-            string title = audioStream.tag
-                .Where(x => x.key == "title")
-                .Select(x => x.value)
-                .FirstOrDefault();
-
-            if(title is null)
-            {
-                title = audioStream.index.ToString();
-            }
-            this.title = title;
-            return this;
-        }
-
-        public IAudioCommandBuilder WithCodec()
-        {
-            if (codecSupported)
-            {
-                return this;
-            }
-            commands.Add($"-c:a aac -b:a {audioStream.bit_rate * 1.1}");
-            return this;
-        }
-    }
-
-    internal interface IFFmpegCommandBuilder
-    {
-        FfmpegRenderedCommand Build();
-        IFFmpegCommandBuilder WithVideoCommands(IEnumerable<MediaStream> videoStreams, IEnumerable<IQuality> qualities, int framerate, int keyframeInterval, int defaultBitrate);
-        IFFmpegCommandBuilder WithAudioCommands(IEnumerable<MediaStream> streams);
-        IFFmpegCommandBuilder WithSubtitleCommands(IEnumerable<MediaStream> streams);
-    }
-
-    internal interface IVideoCommandBuilder
-    {
-        StreamVideoFile Build();
-        IVideoCommandBuilder WithSize(IQuality quality);
-        IVideoCommandBuilder WithBitrate(int bitrate);
-        IVideoCommandBuilder WithBitrate(int bitrate, int defaultBitrate);
-        IVideoCommandBuilder WithPreset(H264Preset preset);
-        IVideoCommandBuilder WithProfile(H264Profile profile);
-        IVideoCommandBuilder WithProfileLevel(string level);
-        IVideoCommandBuilder WithPixelFormat(string format);
-        IVideoCommandBuilder WithFramerate(int framerate);
-        IVideoCommandBuilder WithVideoCodec(string sourceCodec, int keyInterval, bool enableCopy);
-    }
-
-    internal interface IAudioCommandBuilder
-    {
-        StreamAudioFile Build();
-        IAudioCommandBuilder WithLanguage();
-        IAudioCommandBuilder WithTitle();
-        IAudioCommandBuilder WithCodec();
     }
 }
